+++ src/index.html
@@ -35,7 +35,7 @@
 
         </div>
 
-        <div id="popup">
+        <div class="popup">
 
         </div>
     </section>
